Calculate muzzle locations as float-vector, for convenience (and because IMDs store info as float anyway).
git-svn-id: svn+ssh://svn.gna.org/svn/warzone/trunk@5121 4a71c877-e1ca-e34f-864e-861f7616d084

+++ src/droid.h
@@ -234,7 +234,7 @@ extern UDWORD	getNumDroidsForLevel(UDWORD	level);
 
 extern BOOL activateGroupAndMove(UDWORD playerNumber, UDWORD groupNumber);
 /* calculate muzzle tip location in 3d world added int weapon_slot to fix the always slot 0 hack*/
-extern BOOL calcDroidMuzzleLocation(DROID *psDroid, Vector3i *muzzle, int weapon_slot);
+extern BOOL calcDroidMuzzleLocation(DROID *psDroid, Vector3f *muzzle, int weapon_slot);
 
 /* gets a template from its name - relies on the name being unique */
 extern DROID_TEMPLATE * getTemplateFromUniqueName(const char *pName, unsigned int player);
